third_party/WebKit/LayoutTests/external/wpt/workers/shared-worker-name-via-options-expected.txt

This is a testharness.js-based test.
FAIL Test the name property of shared workers mixing constructor options and constructor strings Uncaught Error: assert_equals: expected "my name" but got "[object Object]"
Harness: the test ran to completion.
